requirements.txt
requests
streamlit
pandas
plotly
python-dotenv
diskcache